Complete Otto 2 upgrade for web core
The `ControllerHelpers` module in `apps/web/core/controller_helpers.rb`
has been integrated into `apps/web/core/controllers/base.rb`. This
consolidation centralizes common controller logic and improves code
organization.

Key changes include: - Removed the `apps/web/core/controller_helpers.rb`
file. - Included `Onetime::Helpers::SessionHelpers` and
`Onetime::Helpers::ShrimpHelpers` directly in `Core::Controllers::Base`.
- Moved relevant methods such as `authentication_enabled?`,
`capture_error`, and `capture_message` from `ControllerHelpers` to
`Core::Controllers::Base`. - Updated controllers that were including
`Core::ControllerHelpers` to now include `Core::Controllers::Base`.
