+++ apps/web/core/controllers.rb
@@ -1,6 +1,5 @@
 # apps/web/core/controllers.rb
 
-require_relative 'controller_helpers'
 require_relative 'controllers/page'
 require_relative 'controllers/account'
 require_relative 'controllers/welcome'
