+++ apps/web/core/controllers/data_incoming.rb
@@ -8,29 +8,29 @@ module Core
       include Controllers::Base
 
       def create_incoming
-        publically(req.request_path) do
-          if OT.conf['incoming'] && OT.conf['incoming']['enabled']
-            strategy_result = Otto::Security::Authentication::StrategyResult.new(
-              session: session,
-              user: cust,
-              auth_method: 'session',
-              metadata: {
-                ip: req.client_ipaddress,
-                user_agent: req.user_agent
-              }
-            )
+        return res.redirect '/' unless _is_enabled
 
-            logic = V2::Logic::Incoming::CreateIncoming.new strategy_result, req.params, locale
-            logic.raise_concerns
-            logic.process
-            req.params.clear
-            view     = Core::Views::Incoming.new req, session, cust, locale
-            view.add_message view.i18n[:page][:incoming_success_message]
-            res.body = view.render
-          else
-            res.redirect '/'
-          end
-        end
+        strategy_result = Otto::Security::Authentication::StrategyResult.new(
+            session: session,
+            user: cust,
+            auth_method: 'session',
+            metadata: {
+              ip: req.client_ipaddress,
+              user_agent: req.user_agent
+            }
+          )
+
+        logic = V2::Logic::Incoming::CreateIncoming.new strategy_result, req.params, locale
+        logic.raise_concerns
+        logic.process
+        req.params.clear
+        view     = Core::Views::Incoming.new req, session, cust, locale
+        view.add_message view.i18n[:page][:incoming_success_message]
+        res.body = view.render
+      end
+
+      def _is_enabled
+        OT.conf.dig('incoming', 'enabled') || false
       end
     end
   end
